+++ app/src/main/kotlin/xyz/mpv/rex/ui/player/MPVView.kt
@@ -298,6 +298,7 @@ class MPVView(
       "demuxer-cache-duration" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,
       "demuxer-cache-time" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,
       "demuxer-cache-state/cache-end" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,
+      "demuxer-cache-state/reader-pts" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,
       "cache-buffering-state" to MPVLib.MpvFormat.MPV_FORMAT_INT64,
       "audio-delay" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,
       "sub-delay" to MPVLib.MpvFormat.MPV_FORMAT_DOUBLE,